+++ src/pykeen/training/__init__.py
@@ -16,6 +16,7 @@ from typing import Set, Type
 
 from class_resolver import Resolver
 
+from .callbacks import TrainingCallback  # noqa: F401
 from .lcwa import LCWATrainingLoop  # noqa: F401
 from .slcwa import SLCWATrainingLoop  # noqa: F401
 from .training_loop import NonFiniteLossError, TrainingLoop  # noqa: F401
@@ -26,6 +27,7 @@ __all__ = [
     'LCWATrainingLoop',
     'NonFiniteLossError',
     'training_loop_resolver',
+    'TrainingCallback',
 ]
 
 _TRAINING_LOOP_SUFFIX = 'TrainingLoop'